Replace some usages of next/router with next/navigation in main layout
