+++ apps/store/src/utils/l10n/useCurrentLocale.ts
@@ -1,8 +1,18 @@
-import { useRouter } from 'next/router'
+import { useParams, usePathname } from 'next/navigation'
 import { getLocaleOrFallback } from '@/utils/l10n/localeUtils'
 import { RoutingLocale } from '@/utils/l10n/types'
 
 export const useCurrentLocale = () => {
-  const router = useRouter()
-  return getLocaleOrFallback(router.query.locale as RoutingLocale)
+  // useParams() would return undefined during static rendering and hydration of statically rendered pages
+  // But usePathname keeps working, so we're using it as a fallback
+  let { locale } = (useParams() as { locale: RoutingLocale } | undefined) ?? {}
+  // usePathname will return null when building pages without static paths, hence the fallback
+  // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion
+  const pathname = usePathname() as string | null
+  if (locale === undefined) {
+    if (pathname != null) {
+      locale = pathname.split('/')[1] as RoutingLocale
+    }
+  }
+  return getLocaleOrFallback(locale)
 }
