+++ apps/store/src/utils/mergeSearchParams.ts
@@ -1,18 +1,6 @@
-import { ParsedUrlQuery } from 'querystring'
-import { ORIGIN_URL } from './PageLink'
+import * as querystring from 'querystring'
+import { ReadonlyURLSearchParams } from 'next/navigation'
+import { mergeParsedUrlQuery } from '@/utils/mergeParsedUrlQuery'
 
-export const mergeSearchParams = (linkUrl: string, search: ParsedUrlQuery): string => {
-  const url = new URL(linkUrl, ORIGIN_URL)
-
-  Object.entries(search).forEach(([key, value]) => {
-    if (Array.isArray(value)) {
-      value.forEach((item) => url.searchParams.append(key, item))
-    } else if (typeof value === 'string') {
-      url.searchParams.set(key, value)
-    } else {
-      url.searchParams.set(key, '')
-    }
-  })
-
-  return url.href.replace(ORIGIN_URL, '')
-}
+export const mergeSearchParams = (url: string, searchParams: ReadonlyURLSearchParams): string =>
+  mergeParsedUrlQuery(url, querystring.parse(searchParams.toString()))
